refactor: move reservation files to owned directory
