+++ src/main/java/nextstep/domain/reservation/model/Identity.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.domain.reservation.model;
 
 import java.util.HashMap;
 import java.util.Map;
