+++ src/main/java/nextstep/domain/reservation/model/Reservation.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.domain.reservation.model;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
